refactor: extract gpkg point layer builder
