+++ gpkg_layer_builders.py
@@ -5,6 +5,7 @@ This module provides standalone functions that build ``QgsVectorLayer`` objects
 from activity records and atlas plans.  It contains no I/O — callers are
 responsible for writing the returned layers to disk.
 
+The activity-point builder lives in :mod:`gpkg_point_layer_builder`.
 Geometry-less atlas helper-table builders (document summary, cover highlights,
 page detail items, profile samples, TOC) live in
 :mod:`gpkg_atlas_table_builders`.
@@ -22,7 +23,6 @@ from qgis.core import (
 
 from .gpkg_schema import (
     ATLAS_FIELDS,
-    POINT_FIELDS,
     START_FIELDS,
     TRACK_FIELDS,
     make_qgs_fields,
@@ -36,7 +36,7 @@ from .gpkg_atlas_table_builders import (
     build_toc_layer,
 )
 from .atlas.publish_atlas import build_atlas_page_plans
-from .time_utils import add_seconds_iso
+from .gpkg_point_layer_builder import build_point_layer
 
 
 def build_track_layer(records):
@@ -118,62 +118,6 @@ def build_start_layer(records):
         feature["distance_m"] = record.get("distance_m")
         feature["last_synced_at"] = record.get("last_synced_at")
         features.append(feature)
-
-    provider.addFeatures(features)
-    layer.updateExtents()
-    return layer
-
-
-def build_point_layer(records, write_activity_points=False, point_stride=1):
-    """Build and return a memory ``QgsVectorLayer`` of per-point stream data."""
-    layer = QgsVectorLayer("Point?crs=EPSG:4326", "activity_points", "memory")
-    provider = layer.dataProvider()
-    provider.addAttributes(make_qgs_fields(POINT_FIELDS))
-    layer.updateFields()
-
-    features = []
-    if not write_activity_points:
-        provider.addFeatures(features)
-        layer.updateExtents()
-        return layer
-
-    stride = max(1, int(point_stride or 1))
-    for index, record in enumerate(records, start=1):
-        geometry_points = record.get("geometry_points") or []
-        if len(geometry_points) < 1:
-            continue
-
-        stream_metrics = ((record.get("details_json") or {}).get("stream_metrics") or {})
-        sampled_points = _sample_points(geometry_points, stride)
-        total_points = max(1, len(geometry_points) - 1)
-        for point_index, lat, lon in sampled_points:
-            stream_time_s = _metric_value(stream_metrics, "time", point_index, as_int=True)
-            feature = QgsFeature(layer.fields())
-            feature.setGeometry(QgsGeometry.fromPointXY(QgsPointXY(float(lon), float(lat))))
-            feature["activity_fk"] = index
-            feature["source"] = record.get("source")
-            feature["source_activity_id"] = record.get("source_activity_id")
-            feature["point_index"] = point_index
-            feature["point_ratio"] = float(point_index) / float(total_points)
-            feature["stream_time_s"] = stream_time_s
-            feature["point_timestamp_utc"] = add_seconds_iso(record.get("start_date"), stream_time_s)
-            feature["point_timestamp_local"] = add_seconds_iso(record.get("start_date_local"), stream_time_s)
-            feature["stream_distance_m"] = _metric_value(stream_metrics, "distance", point_index)
-            feature["altitude_m"] = _metric_value(stream_metrics, "altitude", point_index)
-            feature["heartrate_bpm"] = _metric_value(stream_metrics, "heartrate", point_index)
-            feature["cadence_rpm"] = _metric_value(stream_metrics, "cadence", point_index)
-            feature["watts"] = _metric_value(stream_metrics, "watts", point_index)
-            feature["velocity_mps"] = _metric_value(stream_metrics, "velocity_smooth", point_index)
-            feature["temp_c"] = _metric_value(stream_metrics, "temp", point_index)
-            feature["grade_smooth_pct"] = _metric_value(stream_metrics, "grade_smooth", point_index)
-            feature["moving"] = _metric_value(stream_metrics, "moving", point_index, as_int=True)
-            feature["name"] = record.get("name")
-            feature["activity_type"] = record.get("activity_type")
-            feature["start_date"] = record.get("start_date")
-            feature["distance_m"] = record.get("distance_m")
-            feature["geometry_source"] = record.get("geometry_source") or "stream"
-            feature["last_synced_at"] = record.get("last_synced_at")
-            features.append(feature)
 
     provider.addFeatures(features)
     layer.updateExtents()
@@ -301,33 +245,3 @@ def _fallback_geometry(record):
     ])
 
 
-def _sample_points(points, stride):
-    if not points:
-        return []
-    if stride <= 1:
-        return [(index, lat, lon) for index, (lat, lon) in enumerate(points)]
-
-    sampled_indexes = list(range(0, len(points), stride))
-    if sampled_indexes[-1] != len(points) - 1:
-        sampled_indexes.append(len(points) - 1)
-    return [(index, points[index][0], points[index][1]) for index in sampled_indexes]
-
-
-def _metric_value(stream_metrics, key, index, as_int=False):
-    values = stream_metrics.get(key) if isinstance(stream_metrics, dict) else None
-    if not isinstance(values, list) or index >= len(values):
-        return None
-    value = values[index]
-    if value is None:
-        return None
-    if isinstance(value, bool):
-        return int(value) if as_int else value
-    if as_int:
-        try:
-            return int(value)
-        except (TypeError, ValueError):
-            return None
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        return None
