+++ gpkg_write_orchestration.py
@@ -18,10 +18,10 @@ but contains no schema definitions or repository logic.
 from .gpkg_io import write_layer_to_gpkg
 from .gpkg_layer_builders import (
     build_atlas_layer,
-    build_point_layer,
     build_start_layer,
     build_track_layer,
 )
+from .gpkg_point_layer_builder import build_point_layer
 from .gpkg_atlas_table_builders import (
     build_cover_highlight_layer,
     build_document_summary_layer,
