+++ sonar-project.properties
@@ -6,5 +6,5 @@ sonar.sources=.
 sonar.tests=tests
 sonar.exclusions=dist/**,.git/**,.github/**,__pycache__/**,*.pyc,tests/**
 sonar.test.inclusions=tests/**/*.py
-sonar.coverage.exclusions=__init__.py,gpkg_atlas_table_builders.py,gpkg_io.py,gpkg_layer_builders.py,gpkg_schema.py,gpkg_write_orchestration.py,gpkg_writer.py,layer_manager.py,layer_style_service.py,qfit_config_dialog.py,qfit_dockwidget.py,qfit_plugin.py,scripts/**/*.py
+sonar.coverage.exclusions=__init__.py,gpkg_atlas_table_builders.py,gpkg_io.py,gpkg_layer_builders.py,gpkg_point_layer_builder.py,gpkg_schema.py,gpkg_write_orchestration.py,gpkg_writer.py,layer_manager.py,layer_style_service.py,qfit_config_dialog.py,qfit_dockwidget.py,qfit_plugin.py,scripts/**/*.py
 sonar.python.coverage.reportPaths=coverage.xml
